requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.2.1
pyjwt==2.8.0
bcrypt==4.1.3
redis[asyncio]==5.0.4
httpx==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
llama-cpp-python==0.2.77
